refactor(bgrec): split Part Event Polling region into BackgroundRecorder.PartEventPolling.cs
